requirements.txt
opencv-python>=4.9.0.80
Pillow>=10.0.0
numpy>=1.26.0

